+++ tests/Database/Adapter/AdapterTest.php
@@ -2,6 +2,7 @@
 
 namespace Phoenix\Tests\Database\Adapter;
 
+use PDO;
 use Phoenix\Database\Adapter\MysqlAdapter;
 use Phoenix\Database\Adapter\PgsqlAdapter;
 use Phoenix\Database\Adapter\SqliteAdapter;
@@ -15,6 +16,9 @@ class AdapterTest extends PHPUnit_Framework_TestCase
         $pdo = new FakePdo();
         $adapter = new MysqlAdapter($pdo);
         $this->assertInstanceOf('\Phoenix\Database\QueryBuilder\QueryBuilderInterface', $adapter->getQueryBuilder());
+
+        $this->setExpectedException('\LogicException', 'Not yet implemented');
+        $adapter->tableInfo('table');
     }
     
     public function testPgsqlAdapter()
@@ -22,12 +26,26 @@ class AdapterTest extends PHPUnit_Framework_TestCase
         $pdo = new FakePdo();
         $adapter = new PgsqlAdapter($pdo);
         $this->assertInstanceOf('\Phoenix\Database\QueryBuilder\QueryBuilderInterface', $adapter->getQueryBuilder());
+
+        $this->setExpectedException('\LogicException', 'Not yet implemented');
+        $adapter->tableInfo('table');
     }
     
     public function testSqliteAdapter()
     {
-        $pdo = new FakePdo();
+        $pdo = new PDO('sqlite::memory:');
         $adapter = new SqliteAdapter($pdo);
         $this->assertInstanceOf('\Phoenix\Database\QueryBuilder\QueryBuilderInterface', $adapter->getQueryBuilder());
+        
+        $pdo->query('CREATE TABLE "test" ("id" integer PRIMARY KEY AUTOINCREMENT NOT NULL,"title" varchar(255) NOT NULL);');
+        
+        $tableInfo = $adapter->tableInfo('test');
+        $this->assertCount(2, $tableInfo);
+        $this->assertArrayHasKey('id', $tableInfo);
+        $this->assertArrayHasKey('title', $tableInfo);
+        
+        foreach ($tableInfo as $column) {
+            $this->assertInstanceOf('\Phoenix\Database\Element\Column', $column);
+        }
     }
 }